pyproject.toml
[tool.poetry]
name = "parallelized_algorithmic_trader"
version = "1.1.0"
description = "A python platform for backtesting trading strategies with polygon data source and TD Ameritrade integration."
authors = ["brucejiverson <brucejamesiverson@gmail.com>"]
license = "MIT"
classifiers = [
   "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]


[tool.poetry.dependencies]
python = ">=3.10.0,<4.0.0"
numpy = ">=1.19.3"
matplotlib = "^3.3.4"
pyqt5 = "^5.15.9"                   # for GUI with matplotlib
scipy  = ">=1.6.1"
pandas-ta = "^0.2.45b0"
polygon-api-client = "^1.7.1"
hyperopt = "^0.2.7"
lxml = "^4.9.2"
alpaca-py = ">=0.7.2"
PyPDF2 = "=3.0.1"


# optional dependencies for examples
quantstats = "^0.0.59"
[tool.poetry.extras]
hyperopt = ["hyperopt"]             # used in an example

# [tool.poetry.scripts]
# example_command = "path.to.file:function_name"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core", "setuptools>=42", "wheel"]
build-backend = "poetry.masonry.api"
